requirements.txt
mkdocs
mkdocs-material
pymdown-extensions
PyYAML
psycopg2-binary

# Опциональные зависимости для будущих задач документации:
# mkdocs-mermaid2-plugin
# mkdocs-git-revision-date-localized-plugin
